Modify render[_anti]_criteria logic
Modify the implementation of LayoutBlock::should_draw so that it may
take into account multiple RenderCriteria and not only take into account
the last instance of RenderCriteria from each vector.

Additionally, apply the following logic in regards to render_criteria
and render_anti_criteria:

In order for a LayoutBlock to render, **ALL** of its render_criteria
should be satisfied, when **NONE** of its render_anti_criteria should.